Fix a crash on exit
If Surge and MTSMaster were loaded into Hosting AU with MTSMaster
"after" surge in the track list, there was a crash on exit. The
crash on exit was caused by Deregister not checking if the master
was online, and HostingAU reaping the MTSMaster before the Surge
instance. Fix it by guarding the deregister.